Recommend port 15353 when running in user space
Summary: Backport of the same change I made to the upstream: https://github.com/sipa/bitcoin-seeder/pull/95

Test Plan:
On some servers I encountered mDNS services I was not aware of. I think it is better to not mislead
users to think port 5353 is a good choice for running in user space.

This command works well for investigation:
  netstat -pln | grep seeder

Reviewers: O1 Bitcoin ABC, #bitcoin_abc, deadalnix

Reviewed By: O1 Bitcoin ABC, #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D10208