[SECP256K1] Use the appropriated debian version on CI
Summary:
The Cirrus CI is expecting Debian Buster, but the Dockerfile is based on the `debian:stable` image which is now Bullseye.
The build can probably be adapted to run on Bullseye but for now this diff just unbreaks the CI.

Test Plan: Look at the Cirrus CI logs and confirm this should solve the issue.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D10192