[SECP256K1] [RFC] Remove OpenSSL testing support
Summary:
```
This removes the ability to test against OpenSSL, as well as the OpenSSL verification benchmark.

The motivation is that OpenSSL 3 is deprecating part of the API used here (see #869), and I'm not sure it's worth maintaining. We do lose the fact that this is the only test that verifies randomly-generated cases against an independent implementation. On the other hand, there are tons of existing fixed tests now that test all kinds of edge cases already.
```

Backport of [[https://github.com/bitcoin-core/secp256k1/pull/983 | secp256k1#983]].

This fixes the issue https://github.com/Bitcoin-ABC/secp256k1/issues/1

Test Plan:
  ninja check-secp256k1

Check there is no remaining use:
  git grep ENABLE_OPENSSL_TESTS

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D11135