[avalanche] Make regular proof format deserialization tests in par with the legacy
Summary:
This clones the legacy test vectors and use them for the regular format check. This makes sure that no regression from the legacy format is introduced. Of course these tests needs to be updated to conform with the new format, so a valid payout script is added to each case and the stakes signatures updated accordingly.

Ref T1854.

Test Plan:
  ninja check-avalanche-proof_tests

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Maniphest Tasks: T1854

Differential Revision: https://reviews.bitcoinabc.org/D10216