symbol-check: Check requested ELF interpreter
Summary:
It is important that binaries request a standard interpreter location
where most distros would place the linker-loader. Otherwise, the user
would be met with a very confusing message:

    bash: <path>/<to>/bitcoind: No such file or directory

When really it's the interpreter that's not found.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/23148 | core#23148]]

Test Plan:
gitian linux
guix build

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13953