test: Delete generate* calls from TestNode
Summary:
```
Deleting the methods is needed for #22567 to pave the way to make it easier to implicitly call the sync_all member function.
```

Backport of [[https://github.com/bitcoin/bitcoin/pull/23207 | core#23207]].

Includes a conversion leftover to interface_zmq, but not the other test changes that don't apply to us.

Test Plan:
  ninja check-functional-extended

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D12521