use the correct rounding in assert_fee_amount
Summary:
The rounding was introduced in D3516 for a good reason (making it easier to troubleshoot failing tests). But after changing the unit to XEC in the tests, the benefits were lost because it was still round to 8 decimal places.

Using `satoshi_round` instead of switching from `round(..., 8)` to `round(..., 2)` makes the intent more clear.

Test Plan: `ninja check-functional`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D10267