[avalanche] Use a consistent and well known proof master in proof tests
Summary:
The private key associated with the proof master for some test vectors is unknown. This makes it impossible to introduce the proof signature without having to change this key, which this diff does. This makes it easier to review the changes when the proof signature is introduced.

Since changing the proof master will change the proof id, the vectors where the verification result is expected to be valid needs to update the stake signature as well.

The review may show a lot of changes for some vectors because they were using an uncompressed pubkey, and the new one is compressed causing the hex to get unaligned.

Ref T1676.

Test Plan:
  ninja check-avalanche-proof_tests

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: PiRK

Maniphest Tasks: T1676

Differential Revision: https://reviews.bitcoinabc.org/D10300