[avalanche] Select if a conflicting proof should be preferred
Summary:
This diff adds a function to select if a conflicting proof should be preferred over an existing one. This function will be used to determine if a conflicting proof should be added to the regular proof pool or the orphan proof in a follow up.

Ref T1854.

Depends on D10377.

Test Plan:
  ninja check-avalanche

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Maniphest Tasks: T1854

Differential Revision: https://reviews.bitcoinabc.org/D10388