Fix signed integer overflow in prioritisetransaction RPC
Summary:
```
Signed integer overflow is UB in theory, but not in practice. Still,
it would be nice to avoid this UB to allow Bitcoin Core to be
compiled with sanitizers such as -ftrapv or ubsan.

It is impossible to predict when and if an overflow occurs, since
the overflow caused by a prioritisetransaction RPC might only be
later hit when descendant txs are added to the mempool.
Since it is impossible to predict reliably, leave it up to the user
to use the RPC endpoint responsibly, considering their mempool
limits and usage patterns.
```

Backport of [[https://github.com/bitcoin/bitcoin/pull/23418 | core#23418]].

Note that we don't have the UBSAN suppression because we did not enable signed integer overflow detection yet.

Depends on D20439.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D20440