[avalanche] Store proofs associated with utxos instead of peer ids
Summary:
We are willing to deal with conflicting proofs, so this structure will make it easier. A conflicting proof (i.e. a proof sharing at least on UTXO with another) might not have a peer associated, e.g. in case of an orphan. There is no change in behavior.

Ref T1854.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Maniphest Tasks: T1854

Differential Revision: https://reviews.bitcoinabc.org/D10436