[secp256k1] ci: Test `make precomp`
Summary:
Backport/cmake note:
I'm not sure how we could add cmake targets similar to `clean-precomp`. If we add a `secp256k1-clean-precomp` target to remove the two header files we then need to rerun `cmake` to detect that the header files are missing and that we need to conditionnaly  generate the `gen_ecmult_gen_static_prec_table` and `gen_ecmult_static_pre_g` targets and make them dependencies of the `secp256k1` target.

For now we need to manually delete the header files to automatically regenerate them with cmake. Or in the case of ecmult_static_pre_g.h, we do it automatically also if SECP256K1_ECMULT_WINDOW_SIZE is > 15.

This is a partial backport of [[https://github.com/bitcoin-core/secp256k1/pull/988 | secp256k1#988]]
https://github.com/bitcoin-core/secp256k1/pull/988/commits/bb36fe9be0998c81ebc9f18e122bb7617d919877
Depends on D19400

Test Plan:
land and check cirrus build

```
ninja clean-precomp
ninja gen-precomp
git status
```

Change one of the precomputed files, rerun gen-precomp, check that it is restored

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19401