test: refactor: use `create_block` parameters
Summary:
Passing a list of transactions `txlist` to `create_block` appends
them to the block, hence we don't need to do that manually anymore.
The merkle root calculation can also be removed, since this is done
in the end of the helper.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/23521 | core#23521]]
Notes:
- We need to specify version=4 explicitly (missing backport: [[https://github.com/bitcoin/bitcoin/pull/16333 | core#16333]])
- The `txlist` parameter to create_block is from `core#19401`, it was skipped in D10688 (because unused at the time).
- We also move the tx ctor sorting code to `create_block`

Test Plan: `ninja check-functional-extended`

Reviewers: #bitcoin_abc, roqqit

Reviewed By: #bitcoin_abc, roqqit

Differential Revision: https://reviews.bitcoinabc.org/D17691