[avalanche] Order the peer set by score
Summary:
This diff adds an ordered view to the peer set so that highest scores are placed first. This will be useful for selecting the peers we want to get nodes for.

Another diff will cache the score so it's not recomputed every time when it's not needed.

Ref T1864.

Test Plan:
  ninja check-avalanche

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Maniphest Tasks: T1696, T1864

Differential Revision: https://reviews.bitcoinabc.org/D10031