Remove GetSpendHeight
Summary:
It is unclear what the goal of the helper is, as the caller already knows the spend height before calling the helper.

Also, in case the coins view is corrupted, LookupBlockIndex will return nullptr. Dereferencing a nullptr is UB.

Fix both issues by removing it. Also, add a sanity check, which aborts if the coins view is corrupted.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/23630 | core#23630]]

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D12455