[avalanche] Don't create invalid proofs in the conflicting_proof_selection test
Summary: There is a typo in the `buildProofFromAmounts` lambda, and the coin added to the utxo set can have the wrong value. This has no impact on the test now because the proof is not tested for validity, but it would cause an amount mismatch error if doing so.

Test Plan:
  ninja check-avalanche

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D10515