[avalanche] Introduce the ProofPool multi index structure
Summary:
This diff adds a generic multi index container for storing the proofs, with an utxo based mapping. The proofs can then be retrieve by utxos but also by proofid as needed. This container will be used to replace the orphan pool in a follow-up, hence the name validProofPool by opposition to the coming orphanProofPool. There is no change in behavior.

Ref T1854.

Depends on D10520.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Maniphest Tasks: T1854

Differential Revision: https://reviews.bitcoinabc.org/D10521