[avalanche] Add wrappers to add and remove a proof from a proof pool
Summary:
This will make it easier to use the ProofPool structure to store the orphans. It makes the ProofPool a class, but the internal container is kept public to avoid bloating the diff with more changes. Accessors can be added as needed later to make it private. There is no change in behavior.

Ref T1854.

Depends on D10582.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Maniphest Tasks: T1854

Differential Revision: https://reviews.bitcoinabc.org/D10522