[Cashtab] Bumping extension version and synchronizing styles
Summary:
Bumping extension version to get latest OP_RETURN and eToken features available in the extension. This diff required adding some css styling to the extension version of App.js to match the web wallet version.

Known issue in that extension icon creation requires the user to run the extension in its own popped out tab. Acceptable because extension use continues to be much lower than web wallet use and not a breaking change.

Recently landed biometric feature is not active in the extension (wrapping component is not included).

Test Plan:
`npm run extension`
Run the extension locally. Open in new tab. Upload an eToken icon (but don't create the token, the dev version of the extension will not be accepted by the backend). Confirm styling is presentable.

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D10596