[secp256k1] build: Remove #undef hack for ASM in the precomputation programs
Summary:
This was necessary because we used to cross-compile the library but
compile the precomputation programs for the build host. Now it's no
longer necessary and we can cleanly link even the external ASM
(which was the intent of #935).

On the way, remove an obsolete "-I" parameter.

This is a partial backport of [[https://github.com/bitcoin-core/secp256k1/pull/988 | secp256k1#988]]
https://github.com/bitcoin-core/secp256k1/pull/988/commits/7dfceceea692c4118829c06128c41623b2373ac2

Depends on D19401

Test Plan:
delete the pregenerated headers, then
`ninja sepc256k1`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19402