[Cashtab] Change notification position on mobile
Summary: Transaction notifications were not legible in some mobile views. Function added to notifications.js to detect whether user is on mobile or desktop, and sets the notification's style accordingly. On mobile, notification now has margin-top: 10% to make it legible if the device is a mobile.

Test Plan:
cd web/cashtab && npm start
Inspect and toggle device to mobile in chrome
send XEC and ensure notification is pushed lower in the viewport
create token and ensure notification is pushed lower in viewport
submit token icon and ensure notification is pushed lower in viewport
send token and ensure notification is pushed lower in viewport
receive XEC and ensure notification is pushed lower in viewport
receive eToken and ensure notification is pushed lower in viewport
sign message and ensure notification is pushed lower in viewport
force error and ensure notification is pushed lower in viewport

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D10646