[avalanche] Create a conflicting proof pool
Summary:
This moves away from the orphan pool the proofs that are valid and pending due to conflicts.
Since the orphans were previously rescanned, the conflicting proofs needs to do the same, but rather than looking for all the pool we only attempt to register the ones that conflicted with removed peers. The conflicting proof rescan test is updated accordingly.

Ref T1854.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Maniphest Tasks: T1854

Differential Revision: https://reviews.bitcoinabc.org/D10658