[avalanche] Select the favorite orphan in case of a conflict
Summary:
If there is a conflict in the orphan pool, there is no reason to not store the best candidate.
This diff only applies the selection for the proofs added via the registration method. The case of valid proofs being orphaned due to a UTXO being spent will be covered in a follow-up.

Ref T1854.

Depends on D10651 and D10658.

Test Plan:
  ninja all check-avalanche

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Maniphest Tasks: T1854

Differential Revision: https://reviews.bitcoinabc.org/D10524