[avalanche] Add a mode to the proof registration to force accept it as a peer
Summary:
This diff adds a new mode parameter to the proof registration function. This allows for reusing most of the registration code and add the specific behavior that will be requires to account for the votes. There is no change in behavior by default. The new force accept mode will be used when the network accepts the proof during a vote.

Ref T1854.

Test Plan:
  ninja all check-avalanche

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Maniphest Tasks: T1854

Differential Revision: https://reviews.bitcoinabc.org/D10704